Updated values to be integers when rendering

+++ Quartett/Quartett.Web/Models/Characteristic.cs
@@ -3,6 +3,6 @@
     public sealed class Characteristic
     {
         public string Name { get; set; }
-        public double Value { get; set; }
+        public int Value { get; set; }
     }
 }
